Auth

+++ backend/Controllers/Users.php
@@ -14,10 +14,17 @@ class Users extends AbstractController
 
 	public function index() : string
 	{
+		if(!\App::IsAuth()) {
+			return $this->jsonResponse([
+				'message' => 'not_auth'
+			]);
+		}
+
 		$users = $this->usersRepository->getUsers();
 		$response = [];
 		foreach ($users as $user) {
 			$response[] =  [
+				'message' => 'ok',
 				'id' => $user->id(),
 				'name' => $user->name(),
 				'email' => $user->email(),
@@ -30,6 +37,12 @@ class Users extends AbstractController
 
 	public function create() : string
 	{
+		if(!\App::IsAuth()) {
+			return $this->jsonResponse([
+				'message' => 'not_auth'
+			]);
+		}
+
 		$data = $this->jsonRequest();
 		$user = new User(
 			0,
@@ -39,15 +52,25 @@ class Users extends AbstractController
 		);
 		$id = $this->usersRepository->add($user);
 
-		$response = ['id' => $id];
+		$response = [
+			'message' => 'ok',
+			'id' => $id
+		];
 		return $this->jsonResponse($response);
 	}
 
 	public function getById() : string
 	{
+		if(!\App::IsAuth()) {
+			return $this->jsonResponse([
+				'message' => 'not_auth'
+			]);
+		}
+
 		$data = $this->jsonRequest();
 		$user = $this->usersRepository->getById($data['id']);
 		$response =  [
+			'message' => 'ok',
 			'id' => $user->id(),
 			'name' => $user->name(),
 			'email' => $user->email(),
@@ -59,6 +82,12 @@ class Users extends AbstractController
 
 	public function update() : string
 	{
+		if(!\App::IsAuth()) {
+			return $this->jsonResponse([
+				'message' => 'not_auth'
+			]);
+		}
+
 		$data = $this->jsonRequest();
 		$user = new User(
 			$data['id'],
@@ -69,18 +98,28 @@ class Users extends AbstractController
 
 		$id = $this->usersRepository->update($user);
 
-		$response = ['id' => $data['id']];
+		$response = [
+			'message' => 'ok',
+			'id' => $data['id']
+		];
 		return $this->jsonResponse($response);
 	}
 
 	public function delete() : string
 	{
+		if(!\App::IsAuth()) {
+			return $this->jsonResponse([
+				'message' => 'not_auth'
+			]);
+		}
+
 		$data = $this->jsonRequest();
 
 		foreach ($data as $id) {
 			$id = $this->usersRepository->deleteById($id); //TODO needs optimization
 		}
 
+		$data['message'] = 'ok';
 		return $this->jsonResponse($data);
 	}
 }
